Separate sphinx.transforms to multiple files
